line chart: add utility for smoothing line chart
This change ports logic in the vz_line_chart for our consumption in new
GpuLineChart.

Future integration note: unlike vz_line_chart, smoothing is not built-in
to the line chart; previous pattern made the line chart more logic-ful
and made it harder to adopt other smoothing strategies or data
transformation logic. Instead, we decided to have a notion of `aux` or
auxiliary data series and let the consumer pass the smoothed data
series.